abstract calculation test
