+++ tests/lib/test_aggregations.py
@@ -1,0 +1,7 @@
+from test_calculator import TestCalculator
+
+
+class TestAggregations(TestCalculator):
+
+    def setUp(self):
+        TestCalculator.setUp(self)
